Extract getAll functions.
